requirements.txt
fastapi==0.115.8
streamlit==1.40.1
uvicorn==0.34.0
selectolax==0.3.21
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
httpx==0.28.1
mistralai>=1.8.0
pydantic-settings==2.7.1
